requirements.txt
PyVCF8
requests
